@@ -26,7 +26,7 @@
     "build": "npm run bundle && webpack -p",
     "build:umd": "rollup -w -c build/config.js --environment TARGET:web-umd-dev",
     "build:amd": "webpack -p",
-    "lint": "eslint src/apexcharts.js src/**/*.js tests/**/*.js --no-fix",
+    "lint": "eslint src/apexcharts.js src/**/*.js tests/**/*.js",
     "test": "jest --testTimeout=30000",
     "unit": "jest tests/unit/",
     "e2e": "jest tests/e2e/"
@@ -58,15 +58,15 @@
     "concurrently": "^3.5.1",
     "cross-env": "5.2.0",
     "css-loader": "3.2.0",
-    "eslint": "5.14.1",
-    "eslint-config-prettier": "3.3.0",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-import": "2.14.0",
-    "eslint-plugin-prettier": "3.0.1",
-    "eslint-plugin-promise": "4.0.1",
+    "eslint": "6.7.2",
+    "eslint-config-prettier": "6.7.0",
+    "eslint-loader": "2.2.1",
+    "eslint-plugin-import": "2.19.1",
+    "eslint-plugin-prettier": "3.1.2",
+    "eslint-plugin-promise": "4.2.1",
     "jest": "24.1.0",
     "jest-puppeteer": "3.8.0",
-    "prettier": "1.16.4",
+    "prettier": "1.19.1",
     "puppeteer": "1.19.0",
     "puppeteer-core": "1.11.0",
     "rollup": "1.1.0",
