@@ -38,7 +38,11 @@ describe('Y-axis with ultra-small values', () => {
       },
       series: [
         {
-          data: [[1553258700000, 1], [1553259000000, 2], [1553261100000, 4]]
+          data: [
+            [1553258700000, 1],
+            [1553259000000, 2],
+            [1553261100000, 4]
+          ]
         }
       ],
       xaxis: {
